Move implementation for the following methods to CollectionUtilities from INotifyCollectionChangedExtensions
- ObserveElementProperty
- ObserveElementObservableProperty
- ObserveElementPropertyChanged
